Move over the LEDImage class
